+++ R/renderExampleRunPlotSingleCrit1d.R
@@ -32,12 +32,7 @@ renderExampleRunPlot1d = function(x, iter,
 
   propose.points = control$propose.points
   infill.crit.id = getMBOInfillCritId(control$infill.crit)
-  # if (control$multifid) {
-  #   infill.crit.id = "mfEI"
-  #   critfun = infillCritMultiFid.external
-  # } else {
-    critfun = control$infill.crit$fun
-  #}
+  critfun = control$infill.crit$fun
 
   # we need to maximize expected improvement
   opt.direction = getMBOInfillCritMultiplier(control$infill.crit)
@@ -113,15 +108,9 @@ renderExampleRunPlot1d = function(x, iter,
   }
 
   if (isNumeric(par.set, include.int = FALSE)) {
-    gg.fun = data.table::setDF(data.table::melt(evals, id.vars = c(getParamIds(opt.path$par.set), if (se) "se" else NULL)))
-
-    if (control$multifid) {
-      #rename .multifid.lvl according to control object
-      repl = paste0(control$multifid.param, "=", control$multifid.lvls)
-      names(repl) = as.character(seq_along(control$multifid.lvls))
-      #gg.fun$.multifid.lvl = plyr::revalue(as.factor(gg.fun$.multifid.lvl), replace = repl)
-      gg.fun$.multifid.lvl = factor(gg.fun$.multifid.lvl, labels = repl)
-    }
+    evals = data.table::setDT(evals)
+    gg.fun = data.table::melt(evals, id.vars = c(getParamIds(opt.path$par.set), if (se) "se" else NULL))
+    gg.fun = data.table::setDF(gg.fun)
 
     if (se) gg.fun$se = gg.fun$se * se.factor
 
@@ -150,28 +139,16 @@ renderExampleRunPlot1d = function(x, iter,
 
     # finally build the ggplot object(s)
     g = ggplot2::ggplot(data = gg.fun)
-    next.aes = ggplot2::aes_string(x = names.x, y = "value", color = "as.factor(.multifid.lvl)", group = "paste(variable,.multifid.lvl)", linetype = "variable")
-    if (!control$multifid) {
-      next.aes = dropNamed(next.aes, c("group","colour"))
-    }
+    next.aes = ggplot2::aes_string(x = names.x, y = "value", linetype = "variable")
     g = g + ggplot2::geom_line(next.aes, size = line.size)
     g = g + ggplot2::facet_grid(pane~., scales = "free")
     if (se && densregion) {
       #FIXME: We might lose transformation information here tr()
-      next.aes = ggplot2::aes_string(x = names.x, ymin = "value-se", ymax = "value+se", group = ".multifid.lvl")
-      if (!control$multifid) {
-        next.aes = dropNamed(next.aes, "group")
-      }
+      next.aes = ggplot2::aes_string(x = names.x, ymin = "value-se", ymax = "value+se")
       g = g + ggplot2::geom_ribbon(data = gg.fun[gg.fun$variable == "yhat", ], next.aes, alpha = 0.2)
     }
     g = g + ggplot2::geom_point(data = gg.points, ggplot2::aes_string(x = names.x, y = name.y, colour = "type", shape = "type"), size = point.size)
-    if (control$multifid) {
-      mf.colors = tail(RColorBrewer::brewer.pal(n = length(control$multifid.lvls)+1, name = "PuBu"), -1)
-      names(mf.colors) = levels(gg.fun$.multifid.lvl)
-    } else {
-      mf.colors = NULL
-    }
-    g = g + ggplot2::scale_colour_manual(values = c(mf.colors, colors), name = "type")
+    g = g + ggplot2::scale_colour_manual(values = colors, name = "type")
     g = g + ggplot2::scale_linetype(name = "type")
 
     if (noisy) {